Removing the warnings on x86_64 by using ptrdiff_t instead of int in CAST_PATH
git-svn-id: http://boxfs.googlecode.com/svn/trunk@75 a81dc690-963a-11de-99ed-9d3e36ffeeb7

+++ boxfs.c
@@ -16,6 +16,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <stddef.h>
 #include <unistd.h>
 #include <fcntl.h>
 #include <dirent.h>
@@ -24,7 +25,7 @@
 
 #include "boxapi.h"
 
-#define CAST_PATH (char*)(int)
+#define CAST_PATH (char*)(ptrdiff_t)
 
 static int box_access(const char *path, int mask)
 {
